+++ components/Experience.tsx
@@ -2,7 +2,7 @@
 
 import { Suspense } from 'react'
 import { Canvas } from '@react-three/fiber'
-import { useGLTF } from '@react-three/drei'
+import { useGLTF, useTexture } from '@react-three/drei'
 import { Ship } from './Ship'
 import { SolarSystem } from './SolarSystem'
 import { Starfield } from './Starfield'
@@ -10,6 +10,9 @@ import { NightSky } from './NightSky'
 import { Effects } from './Effects'
 
 useGLTF.preload('/planets/various_planets.glb')
+useGLTF.preload('/ship/ship.glb')
+useGLTF.preload('/sun/stroming_sun.glb')
+useTexture.preload('/sky/NightSkyHDRI008_8K.jpg')
 
 export function Experience() {
   return (